+++ openbas-front/src/admin/components/components/lessons/categories/questions/CreateLessonsTemplateQuestion.tsx
@@ -1,27 +1,16 @@
 import React, { FunctionComponent, useState } from 'react';
-import { makeStyles } from '@mui/styles';
-import { ListItemButton, ListItemIcon, ListItemText } from '@mui/material';
-import { ControlPointOutlined } from '@mui/icons-material';
 import { useFormatter } from '../../../../../../components/i18n';
 import LessonsTemplateQuestionForm, { LessonsTemplateQuestionInputForm } from './LessonsTemplateQuestionForm';
 import { addLessonsTemplateQuestion } from '../../../../../../actions/Lessons';
 import Drawer from '../../../../../../components/common/Drawer';
 import { useAppDispatch } from '../../../../../../utils/hooks';
 import type { LessonsTemplateCategory } from '../../../../../../utils/api-types';
-import type { Theme } from '../../../../../../components/Theme';
+import ListItemButtonCreate from '../../../../../../components/common/ListItemButtonCreate';
 
 interface Props {
   lessonsTemplateId: string;
   lessonsTemplateCategoryId: string;
 }
-
-const useStyles = makeStyles((theme: Theme) => ({
-  text: {
-    fontSize: 15,
-    color: theme.palette.primary.main,
-    fontWeight: 500,
-  },
-}));
 
 const CreateLessonsTemplateQuestion: FunctionComponent<Props> = ({
   lessonsTemplateId,
@@ -29,7 +18,6 @@ const CreateLessonsTemplateQuestion: FunctionComponent<Props> = ({
 }) => {
   // Standard hooks
   const { t } = useFormatter();
-  const classes = useStyles();
   const dispatch = useAppDispatch();
 
   const [open, setOpen] = useState(false);
@@ -51,19 +39,10 @@ const CreateLessonsTemplateQuestion: FunctionComponent<Props> = ({
   };
   return (
     <>
-      <ListItemButton
-        divider
+      <ListItemButtonCreate
+        title={t('Create a new lessons learned question')}
         onClick={handleOpen}
-        color="primary"
-      >
-        <ListItemIcon color="primary">
-          <ControlPointOutlined color="primary" />
-        </ListItemIcon>
-        <ListItemText
-          primary={t('Create a new lessons learned question')}
-          classes={{ primary: classes.text }}
-        />
-      </ListItemButton>
+      />
       <Drawer
         open={open}
         handleClose={handleClose}
